+++ cmd/web/routes.go
@@ -3,16 +3,25 @@ package main
 import (
 	"net/http"
 
-	"github.com/bmizerany/pat"
 	"github.com/cpwu/nova/pkg/config"
 	"github.com/cpwu/nova/pkg/handlers"
+	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
 )
 
 func routes(app *config.AppConfig) http.Handler {
-	mux := pat.New()
+	// mux := pat.New()
 
-	mux.Get("/", http.HandlerFunc(handlers.Repo.Home))
-	mux.Get("/about", http.HandlerFunc(handlers.Repo.About))
+	// mux.Get("/", http.HandlerFunc(handlers.Repo.Home))
+	// mux.Get("/about", http.HandlerFunc(handlers.Repo.About))
+	mux := chi.NewRouter()
 
+	// Use the middleware.Recoverer middleware to recover from panics and return a 500 Internal Server Error response.
+	mux.Use(middleware.Recoverer)
+	mux.Use(NoSurf)
+	mux.Use(SessionLoad)
+
+	mux.Get("/", handlers.Repo.Home)
+	mux.Get("/about", handlers.Repo.About)
 	return mux
 }