+++ pkg/config/config.go
@@ -3,6 +3,8 @@ package config
 import (
 	"html/template"
 	"log"
+
+	"github.com/alexedwards/scs/v2"
 )
 
 // AppConfig holds the application configuration settings, including a template cache that maps template names to their corresponding parsed templates. This allows for efficient rendering of templates by avoiding the need to parse them on each request.
@@ -10,4 +12,6 @@ type AppConfig struct {
 	UseCache      bool
 	TemplateCache map[string]*template.Template
 	InfoLog       *log.Logger
+	InProduction  bool
+	Session       *scs.SessionManager
 }