+++ pkg/handlers/handlers.go
@@ -30,6 +30,8 @@ func NewHandlers(r *Repository) {
 
 // Home page handler that writes a simple response to the client and handles any potential errors that may occur during the writing process. It also prints the number of bytes written to the console for debugging purposes.
 func (m *Repository) Home(w http.ResponseWriter, r *http.Request) {
+	remoteIP := r.RemoteAddr
+	m.App.Session.Put(r.Context(), "remote_ip", remoteIP)
 	render.RenderTemplate(w, "home.page.tmpl", &models.TemplateData{})
 }
 
@@ -38,6 +40,9 @@ func (m *Repository) About(w http.ResponseWriter, r *http.Request) {
 	// perform some logic
 	stringMap := make(map[string]string)
 	stringMap["test"] = "Hello again"
+
+	remoteIP := m.App.Session.GetString(r.Context(), "remote_ip")
+	stringMap["remote_ip"] = remoteIP
 
 	// send the data to the template
 	render.RenderTemplate(w, "about.page.tmpl", &models.TemplateData{